@@ -1,12 +1,10 @@
 package de.onlineberatung.authenticator;
 
-import static de.onlineberatung.credential.MailOtpCredentialModel.createOtpModel;
 import static java.time.Clock.systemDefaultZone;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -21,8 +19,8 @@ import de.onlineberatung.otp.OtpService;
 import de.onlineberatung.otp.ValidationResult;
 import de.onlineberatung.credential.MailOtpCredentialModel;
 import de.onlineberatung.keycloak_otp_config_spi.keycloakextension.generated.web.model.OtpType;
-import javax.ws.rs.core.MultivaluedHashMap;
-import javax.ws.rs.core.Response;
+import jakarta.ws.rs.core.MultivaluedHashMap;
+import jakarta.ws.rs.core.Response;
 import org.junit.Before;
 import org.junit.Test;
 import org.keycloak.authentication.AuthenticationFlowContext;
