@@ -8,9 +8,9 @@ import static org.mockito.Mockito.when;
 
 import de.onlineberatung.keycloak_otp_config_spi.keycloakextension.generated.web.model.Challenge;
 import de.onlineberatung.keycloak_otp_config_spi.keycloakextension.generated.web.model.OtpType;
+import jakarta.ws.rs.core.MultivaluedHashMap;
+import jakarta.ws.rs.core.Response;
 import java.util.Collections;
-import javax.ws.rs.core.MultivaluedHashMap;
-import javax.ws.rs.core.Response;
 import org.junit.Before;
 import org.junit.Test;
 import org.keycloak.authentication.AuthenticationFlowContext;
@@ -18,7 +18,8 @@ import org.keycloak.authentication.AuthenticationFlowError;
 import org.keycloak.http.HttpRequest;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
-import org.keycloak.models.UserCredentialManager;
+//import org.keycloak.models.UserCredentialManager;
+import org.keycloak.models.SubjectCredentialManager;
 import org.keycloak.models.UserModel;
 import org.keycloak.models.credential.OTPCredentialModel;
 import org.mockito.ArgumentCaptor;
@@ -29,7 +30,7 @@ public class OtpParameterAuthenticatorTest {
   private AuthenticationFlowContext authFlow;
   private UserModel user;
   private RealmModel realm;
-  private UserCredentialManager credentialManager;
+  private SubjectCredentialManager credentialManager;
   private HttpRequest httpRequest;
   private MultivaluedHashMap<String, String> decodedFormParams;
   private KeycloakSession session;
@@ -44,8 +45,8 @@ public class OtpParameterAuthenticatorTest {
     when(authFlow.getUser()).thenReturn(user);
     session = mock(KeycloakSession.class);
     when(authFlow.getSession()).thenReturn(session);
-    credentialManager = mock(UserCredentialManager.class);
-    when(session.userCredentialManager()).thenReturn(credentialManager);
+    credentialManager = mock(SubjectCredentialManager.class);
+    when(user.credentialManager()).thenReturn(credentialManager);
     httpRequest = mock(HttpRequest.class);
     decodedFormParams = new MultivaluedHashMap<>();
     when(httpRequest.getDecodedFormParameters()).thenReturn(decodedFormParams);
@@ -54,7 +55,8 @@ public class OtpParameterAuthenticatorTest {
 
   @Test
   public void isConfigured_should_be_true_if_otp_auth_is_configured_for_user() {
-    when(credentialManager.isConfiguredFor(realm, user, OTPCredentialModel.TYPE)).thenReturn(true);
+    when(user.credentialManager()).thenReturn(credentialManager);
+    when((credentialManager.isConfiguredFor(OTPCredentialModel.TYPE))).thenReturn(true);
 
     var configured = authenticator.configuredFor(session, realm, user);
 
